Implement authorization using policies

+++ app/Http/Controllers/QuestionsController.php
@@ -4,9 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Question;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Auth;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct(){
+        // calling the authentication middleware
+        $this->middleware('auth', ['except' => ['index', 'show ']]);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -75,6 +83,9 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         //
+        if(Auth::user()->cannot('update', $question)){
+            return redirect()->route('questions.index');
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -88,6 +99,9 @@ class QuestionsController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        if(Auth::user()->cannot('update', $question)){
+            abort(403, "Access denied for this user");
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
@@ -101,7 +115,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        //
+        
+        if(Auth::user()->cannot('delete', $question)){
+            abort(403, "Access denied for this user");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', "Your question has been deleted");
     }
